pyproject.toml
[project]
name = "codemind"
version = "0.1.0"
description = "Local, pre-push AI code review orchestrator"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "CodeMind Contributors" }]
keywords = ["git", "code-review", "ai", "pre-push", "cli", "codemind"]

dependencies = [
    "click>=8.0",
    "pyautogui>=0.9",
    "pyperclip>=1.8",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
mcp = [
    "mcp>=1.0.0",
]

[project.scripts]
codemind = "codemind.cli.main:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["codemind*"]
